fix #2 - move entity changes to trait
